Minor changes to code to prepare integration of TD backups.

+++ src/main/java/csironi/ggp/course/MCTStest/MCSSpeedTest.java
@@ -189,8 +189,8 @@ public class MCSSpeedTest {
 	        double nodesPerSecond = -1;
 	        Role playingRole = null;
 	        Move chosenMove = null;
-	        long scoresSum = -1L;
-	        long visits = -1;
+	        double scoresSum = -1.0;
+	        int visits = -1;
 	        double averageScore = -1;
 
 	        // Try to initialize the propnet state machine.
@@ -234,7 +234,7 @@ public class MCSSpeedTest {
 		 	        scoresSum = finalMove.getScoreSum();
 		 	        visits = finalMove.getVisits();
 		 	        if(visits != 0){
-		 	        	averageScore = ((double) scoresSum) / ((double) visits);
+		 	        	averageScore = scoresSum / ((double) visits);
 		 	        }
 		 	        iterations = MCSmanager.getIterations();
 		        	visitedNodes = MCSmanager.getVisitedNodes();
