+++ src/main/java/csironi/ggp/course/MCTStest/ProverMCTSSpeedTest.java
@@ -157,8 +157,8 @@ public class ProverMCTSSpeedTest {
 	        double nodesPerSecond = -1;
 	        Role playingRole = null;
 	        Move chosenMove = null;
-	        long scoresSum = -1L;
-	        long visits = -1;
+	        double scoresSum = -1.0;
+	        int visits = -1;
 	        double averageScore = -1;
 
 	        // Try to initialize the propnet state machine.
@@ -232,7 +232,7 @@ public class ProverMCTSSpeedTest {
 			    scoresSum = finalMove.getScoreSum();
 			    visits = finalMove.getVisits();
 			    if(visits != 0){
-			    	averageScore = ((double) scoresSum) / ((double) visits);
+			    	averageScore = scoresSum / ((double) visits);
 			    }
 			    iterations = MCTSmanager.getIterations();
 				visitedNodes = MCTSmanager.getVisitedNodes();
