+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCS/manager/prover/ProverMCSManager.java
@@ -131,7 +131,7 @@ public class ProverMCSManager {
 
 				//System.out.println("Visits: " + visits);
 
-				int scoreSum = this.currentMovesStatistics[i].getScoreSum();
+				double scoreSum = this.currentMovesStatistics[i].getScoreSum();
 
 				//System.out.println("Score sum: " + scoreSum);
 
@@ -143,7 +143,7 @@ public class ProverMCSManager {
 
 				}else{
 					// Compute average score
-					currentAvgScore = ((double) scoreSum) / ((double) visits);
+					currentAvgScore = scoreSum / ((double) visits);
 
 					//System.out.println("Computed average score: " + currentAvgScore);
 				}
