+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCS/prover/ProverMcsGamer.java
@@ -153,7 +153,7 @@ public class ProverMcsGamer extends ProverGamer {
     	double iterationsPerSecond = -1;
     	double nodesPerSecond = -1;
     	Move theMove = null;
-    	int moveScoreSum = -1;
+    	double moveScoreSum = -1.0;
     	int moveVisits = -1;
     	double moveAvgScore = -1;
