+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/propnet/strategies/playout/jointmoveselector/MASTJointMoveSelector.java
@@ -54,7 +54,7 @@ public class MASTJointMoveSelector implements JointMoveSelector {
 			moveStats = this.mastStatistics.get(move);
 
 			if(moveStats != null && moveStats.getVisits() != 0){
-				currentAvgScore = ((double) moveStats.getScoreSum()) / ((double) moveStats.getVisits());
+				currentAvgScore = moveStats.getScoreSum() / ((double) moveStats.getVisits());
 			}else{
 				currentAvgScore = 100;
 			}
