+++ src/main/java/org/ggp/base/player/gamer/statemachine/MCTS/manager/prover/strategies/movechoice/ProverMaximumScoreChoice.java
@@ -64,7 +64,7 @@ public class ProverMaximumScoreChoice implements ProverMoveChoiceStrategy {
 
 			//System.out.println("Visits: " + visits);
 
-			int scoreSum = myMovesStats[i].getScoreSum();
+			double scoreSum = myMovesStats[i].getScoreSum();
 
 			//System.out.println("Score sum: " + scoreSum);
 
@@ -74,7 +74,7 @@ public class ProverMaximumScoreChoice implements ProverMoveChoiceStrategy {
 			 * TODO: remove this check once you are reasonably sure that
 			 * this can never happen.
 			 */
-			if(visits < 0 || scoreSum < 0){
+			if(visits < 0 || scoreSum < 0.0){
 				throw new RuntimeException("Negative value for visits and/or scores sum : VISITS=" + visits + ", SCORE_SUM=" + scoreSum + ".");
 			}
 
@@ -86,7 +86,7 @@ public class ProverMaximumScoreChoice implements ProverMoveChoiceStrategy {
 
 			}else{
 				// Compute average score
-				currentAvgScore = ((double) scoreSum) / ((double) visits);
+				currentAvgScore = scoreSum / ((double) visits);
 
 				//System.out.println("Computed average score: " + currentAvgScore);
 			}
